+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+3.13.2
+----------
+`2026-06-02 · 1 🐛`
+
+### 🐛 Bug fixes
+- fix(templates): fix broken WhatsApp template UI
+
+3.13.1
+----------
+`2026-05-29 · 1 🐛`
+
+### 🐛 Bug fixes
+- fix(nav): use named routes in header tabs to prevent incorrect navigation from configured app routes
+
+3.13.0
+----------
+`2026-05-28 · 1 🎉`
+
+### 🎉 New features
+- Introduce WhatsApp Product Led Business Verification Configuration Tab
+
+3.12.1
+----------
+`2026-05-28 · 1 🐛`
+
+### 🐛 Bug fixes
+- fix(gmail): handle OAuth callback postMessage for Module Federation
+
 3.12.0
 ----------
 `2026-05-07 · 1 🔧 · 1 🎉`